move python_component and _component decorator in _component file
