+++ sdk/python/tests/dsl/component_tests.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-from kfp.dsl._python_component import component
+from kfp.dsl._component import component
 from kfp.dsl._types import GCSPath, Integer
 import unittest
